+++ src/services/ElasticsearchAdapter.js
@@ -31,6 +31,10 @@ export default class ElasticsearchAdapter {
     return this.client.catIndices(params, filter)
   }
 
+  catIndexTemplates (params, filter) {
+    return this.client.catIndexTemplates(params, filter)
+  }
+
   catShards (params, filter) {
     return this.client.catShards(params, filter)
   }